refactor to int, attempt at visitor implementation 2

+++ src/main/java/com/designpatterns/paint/base/Models/Shapes/Shape/IShape.java
@@ -14,10 +14,10 @@ public interface IShape
     void drawContour(Graphics g, Color color);
     void setSelected(boolean bool);
     boolean isSelected();
-    double getX();
-    double getWidth();
-    double getY();
-    double getHeight();
+    int getX();
+    int getWidth();
+    int getY();
+    int getHeight();
     void setSize(int newWidth, int newHeight);
     ShapeType getType();
     void setMovingPosition(Double mousePositionX, Double mousePositionY, Double cursorSelectedX, Double cursorSelectedY);
